Odds Ratio Table and Plot

Function for Odds Ratio Table and Plot.

Usage

oddsratio(
  data,
  explanatory,
  outcome,
  outcomeLevel,
  showNomogram = FALSE,
  showExplanations = FALSE
)

Arguments

data

The data as a data frame.

explanatory

The explanatory variables to be used in the analysis.

outcome

The outcome variable to be used in the analysis.

outcomeLevel

Specify which outcome level represents the positive case for likelihood ratio calculations. If not specified, the function will use the second level alphabetically.

showNomogram

Display an interactive nomogram for converting pre-test to post-test probabilities using likelihood ratios calculated from the data.

showExplanations

Display educational explanations for each analysis type to help interpret odds ratios, risk ratios, diagnostic test performance, ROC analysis, and likelihood ratios.

Value

A results object containing:

results$todo					a html

results$text					a html

results$text2					a html

results$plot					an image

results$plot_nomogram					an image

results$nomogram					a html

results$oddsRatioExplanation					a html

results$riskMeasuresExplanation					a html

results$diagnosticTestExplanation					a html

results$nomogramAnalysisExplanation					a html

Examples

# \donttest{
# example will be added
# }